+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+## [2.6.0] - 2021-05-03
+
 ### Added
 
 - Add a new authentication backend `fonzie`
@@ -1295,7 +1297,8 @@ us:
 - finish integrating the missing pages and improve the sandbox environment;
 - test and polish the use of richie as a django app / node dependency.
 
-[unreleased]: https://github.com/openfun/richie/compare/v2.5.0...master
+[unreleased]: https://github.com/openfun/richie/compare/v2.6.0...master
+[2.6.0]: https://github.com/openfun/richie/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/openfun/richie/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/openfun/richie/compare/v2.3.3...v2.4.0
 [2.3.3]: https://github.com/openfun/richie/compare/v2.3.2...v2.3.3